// daweb_30643_mitroi_bianca_assignment3/frontend/src/interfaces/ISectionContent.tsx
export interface SectionContent {
  booking?: boolean;
  id?: number;
  available?: boolean;
  index?: number;
  type?: string;
  title?: string;
  description?: string;
  image?: File;
  location?: string;
  pricePerNight?: number;
  capacity?: number;
  offer?: number;
  link?: string;
  rating?: number;
  user?: string;
}